requirements.txt
python-chess
torch
numpy
streamlit
matplotlib
